// src/app/report1/models/usrrights.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class UsrRights {

    user_name: string;
    branch_code: string;
    menu_name: string;
    rights_company: string;
    rights_admin: string;
    rights_add: string;
    rights_edit: string;
    rights_delete: string;
    rights_print: string;
    rights_email: string;
    rights_docs: string;
    rights_docs_upload: string;
    rights_view: string;
    rights_approval: string;
    user_locked: string;

    _globalvariables: GlobalVariables;
}